And now with fewer parentheses.

+++ src/java/com/threerings/msoy/item/server/persist/ItemRepository.java
@@ -778,8 +778,7 @@ public abstract class ItemRepository<
                  new FromOverride(getRatingClass()),
                  new Where(getRatingColumn(RatingRecord.ITEM_ID), itemId));
 
-        float newRating = 
-            (float) ((average.count == 0) ? 0.0 : ((float) average.sum)/average.count);
+        float newRating = (average.count == 0) ? 0f : average.sum/(float)average.count;
         // and then smack the new value into the item using yummy depot code
         updatePartial(getItemClass(), itemId, ItemRecord.RATING, newRating,
                       ItemRecord.LAST_TOUCHED, new Timestamp(System.currentTimeMillis()));
